requirement.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==2.0.7
django-celery==3.2.2
django-ckeditor==5.5.0
django-js-asset==1.1.0
kombu==3.0.37
Pillow==5.2.0
pytz==2018.5
redis==2.10.6
